Remove queries.py in favour of sparql.py
A supporting file sparql.py was introduced early on to allow us to write
sparql queries.  Queries were added into that file because hackathon.
The mechanism for defining queries is already offered by the template
however.  At this point sparql.py is clean enough as it is and we can
just remove queries.py.  The name queries.py would be more logical but
coming from an AI context and having learned a little bit of LD, the
name sparql.py will likely be the most obvoious "this is the LD bit"
suggestion.