Increase sleep time
When booting up the stack for the first time the initial ingest may fail
when the database is not there yet.  We generally respond to delta
messages and use retries for this case but in this hackathon setting I'm
just adding a timeout.  Not the best but probably not the worst either.